Update reflection code to be compatible with Idris2#2535
[Idris2#2535](github.com/idris-lang/Idris2/pull/2535) adds a quantity
argument to PCase and ICase. Update the reflection code to include
this additional argument.

The fix in `src/Language/Reflection/Syntax.idr` is only temporary, I
think the correct fix there should be a `Nothing` on the quantity,
or alternatively fix downstream from it, requiring a quantity
annotation. I'm using the easy fix first, but might update this PR
with the downstream fix once I have a look around the code-base. (Or
if someone else who knows the codebase wants to do it, I'm happy to
take a step back!)

Don't merge this commit into main until #2535 is merged and
incorproated into the nightly build.